Cast bytesRead to int64_t

+++ src/Azure/Impl/PageBlob.cpp
@@ -76,7 +76,7 @@ namespace AVEVA::RocksDB::Plugin::Azure::Impl
 
         auto bytesRead = content.BodyStream->ReadToCount(reinterpret_cast<uint8_t*>(buffer.data()), buffer.size());
 
-        assert(content.ContentRange.Length.ValueOr(0) == bytesRead && "Bytes read differ from server ContentRange");
+        assert(content.ContentRange.Length.ValueOr(0) == static_cast<int64_t>(bytesRead) && "Bytes read differ from server ContentRange");
         return static_cast<int64_t>(bytesRead);
     }
 }
